Move PersistReturns queries into the module
